Added files via upload
Tried to commit in Netbeans and then push, it gave me an error. Here is the zip file for the project that contains the web pages. All you need to do is create a new empty project, import the zip, and go into the test packages in netbeans, csvData, and right click on one of the .csv files to get to Properties and find the path to update "FillTables.java". Also update dbconfig.properties for your MySQL. Finally, you must go into workbench, delete your containers schema, and run the ContainersSchema.sql in "Website Resources" on Drive. You may have to modify line 1. Now copy query2 and run that in Workbench as well (This will create the db for login).

After doing all that you can run the project, username = admin, password = password